♻️ refactor : 댓글 기능 테스트 코드 리팩토링 및 파일 구조 정리
