+++ apps/posts/tests/comment/test_comment_permissions.py
@@ -8,11 +8,12 @@ User = get_user_model()
 
 class CommentPermissionTests(TestCase):
     def setUp(self) -> None:
-        # 테스트용 유저를 생성합니다.
+        """테스트용 유저 생성"""
         self.user = User.objects.create_user(email="test@example.com", password="testpass", nickname="user")
 
     def test_safe_methods(self) -> None:
-        # SAFE_METHODS(GET 등)에서는 항상 True 반환
+        """SAFE_METHODS(GET 등)에서는 항상 True 반환"""
+
         class DummyRequest:
             method = "GET"
 
@@ -25,7 +26,8 @@ class CommentPermissionTests(TestCase):
         self.assertTrue(perm.has_object_permission(request, None, obj))
 
     def test_not_author(self) -> None:
-        # 작성자가 아닌 경우 PUT/DELETE 등에서는 False 반환
+        """작성자가 아닌 경우 PUT/DELETE 등에서는 False 반환"""
+
         class DummyRequest:
             method = "PUT"
 
@@ -38,7 +40,8 @@ class CommentPermissionTests(TestCase):
         self.assertFalse(perm.has_object_permission(request, None, obj))
 
     def test_author(self) -> None:
-        # 작성자인 경우 PUT/DELETE 등에서는 True 반환
+        """작성자인 경우 PUT/DELETE 등에서는 True 반환"""
+
         class DummyRequest:
             method = "DELETE"
